+++ example/lib/main.dart
@@ -11,6 +11,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData(
+        
         // This is the theme of your application.
         //
         // Try running your application with "flutter run". You'll see the