+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.0]
+
+* Added a Readme
+* Fix initialize splash issue
+* Fix android manifest app name issue
+
 ## [0.0.1]
 
 * Initial version, support Android  only.